Improve intro and audio
Checked for debug keys and disabled them through editor
Added option for embedded sfx to play if no valid external location is picked
Continued title screen

+++ Assets/scripts/ugbm.cs
@@ -5,17 +5,24 @@ using System.IO;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.Networking;
+using UnityEngine.UI;
 
 public class ugbm : MonoBehaviour
 {
 
+    public string resourceFolder = "_FX/BMX"; // Folder name in Resources
+    //private AudioSource audioSource;
+    private AudioClip[] audioClips;
+
+
     public AudioSource audioSource;
-    public string musicDirectory = "C:/Projects/music"; // Change this to the directory where your music files are stored
+    public string musicDirectory = null; //C:/Projects/music  // Change this to the directory where your music files are stored
     string[] musicFiles;
     int audioIndex = 0;
     string verified_directory = "";
     void Start()
     {
+        
         string path = Application.persistentDataPath + "/musicdir.txt";
         if (File.Exists(path))
         {
@@ -32,44 +39,87 @@ public class ugbm : MonoBehaviour
            //reader.Close();
 
             Debug.Log("File content: " + musicDirectory);
+            if (musicDirectory!="")
+            {
+                musicLoader();
+            }
+            else
+            {
+                Debug.Log("MUSIC_INTERNAL");
+                audioSource = gameObject.AddComponent<AudioSource>();
+                audioClips = Resources.LoadAll<AudioClip>(resourceFolder);
+                //  PlayRandomSound();
+            }
+            
         }
         else
         {
             Debug.Log("File not found at: " + path);
+
         }
 
 
-        musicLoader();
+       
 
     }
+
+    void PlayRandomSound()
+    {
+
+    }
+
+
     public void musicLoader()
     {
-        audioSource = this.gameObject.GetComponent<AudioSource>();
-        musicFiles = Directory.GetFiles(musicDirectory, "*.mp3");
-        if (musicFiles.Length > 0)
+        if (musicDirectory != verified_directory)
         {
-            if (musicDirectory!=verified_directory)
+
+            if (audioSource != null)
             {
-                musicFiles = musicFiles.OrderBy(x => Guid.NewGuid()).ToArray();
-                StartCoroutine(PlayMusic2(musicFiles[0])); // Play the first music file found
-
-                string path = Application.persistentDataPath + "/musicdir.txt";
-
-                // Write some text to the file
-                StreamWriter writer = new StreamWriter(path, false);
-                writer.WriteLine(musicDirectory);
-                writer.Close();
-
-                Debug.Log("File written to: " + path);
-
-
-                verified_directory = musicDirectory;
+                if (audioSource.isPlaying)
+                {
+                    audioSource.Stop();
+                }
             }
 
-        }
-        else
-        {
-            verified_directory = "";
+            if (musicDirectory != "" && Directory.Exists(musicDirectory))
+            {
+                audioSource = this.gameObject.GetComponent<AudioSource>();
+                musicFiles = Directory.GetFiles(musicDirectory, "*.mp3");
+                if (musicFiles.Length > 0)
+                {
+                    if (musicDirectory != verified_directory)
+                    {
+                        musicFiles = musicFiles.OrderBy(x => Guid.NewGuid()).ToArray();
+                        StartCoroutine(PlayMusic2(musicFiles[0])); // Play the first music file found
+
+                        string path = Application.persistentDataPath + "/musicdir.txt";
+
+                        // Write some text to the file
+                        StreamWriter writer = new StreamWriter(path, false);
+                        writer.WriteLine(musicDirectory);
+                        writer.Close();
+
+                        Debug.Log("File written to: " + path);
+
+
+                        verified_directory = musicDirectory;
+                    }
+
+                }
+                else
+                {
+                    verified_directory = "";
+                }
+            }
+            else
+            {
+                musicDirectory = "";
+                verified_directory = "";
+                Debug.Log("MUSIC_INTERNAL");
+                audioSource = gameObject.AddComponent<AudioSource>();
+                audioClips = Resources.LoadAll<AudioClip>(resourceFolder);
+            }
         }
     }
     /* IEnumerator PlayMusic(string filePath)
@@ -82,19 +132,53 @@ public class ugbm : MonoBehaviour
          audioSource.Play();
      }
     */
-    void Update()
+
+
+    void LateUpdate()
     {
-        if (audioSource != null && !audioSource.isPlaying)
+
+        resumeMusicCheck();
+
+    }
+
+
+
+
+    void resumeMusicCheck()
+    {
+        
+        //4-2-2025 this will prevent quickly loading a new track
+        if (Application.isFocused && GameObject.Find("txt_Pause")==null)
         {
-            audioIndex++;
-            StartCoroutine(PlayMusic2(musicFiles[audioIndex]));
-            if (audioIndex>musicFiles.Length-1)
+            if (audioSource != null && !audioSource.isPlaying && musicDirectory != "")
             {
-                audioIndex = 0;
+                audioIndex++;
+                StartCoroutine(PlayMusic2(musicFiles[audioIndex]));
+                if (audioIndex > musicFiles.Length - 1)
+                {
+                    audioIndex = 0;
+                }
+            }
+            else
+            {
+                if (!audioSource.isPlaying)
+                {
+                    //No custom music, use the one in resources
+                    if (audioClips.Length == 0)
+                    {
+                        Debug.LogWarning("No audio clips found in the specified folder.");
+                        return;
+                    }
+
+                    AudioClip clip = audioClips[UnityEngine.Random.Range(0, audioClips.Length)];
+                    audioSource.clip = clip;
+                    audioSource.Play();
+                    Invoke(nameof(PlayRandomSound), clip.length);
+                }
+
             }
         }
     }
-
 
     IEnumerator PlayMusic2(string filePath)
     {
@@ -110,7 +194,6 @@ public class ugbm : MonoBehaviour
         else
         {
             Debug.LogError("Error loading audio file: " + request.error);
-
         }
     }
 }
